requirements.txt
Django>=5.2.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
python-jose>=3.3.0
django-cors-headers>=4.7.0
requests>=2.31.0
